requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.1.7
flask==2.2.5
Flask-Migrate==4.0.5
flask-sqlalchemy==3.0.5
greenlet==3.0.0
gunicorn==21.2.0
importlib-metadata==6.5.0
importlib-resources==5.6.0
Mako==1.1.5
MarkupSafe==2.1.3
migrate==0.3.6
packaging==20.9
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==0.19.1
six==1.16.0
SQLAlchemy==1.4.27
werkzeug==2.2.3
zipp==3.4.1
flask-cors==3.0.8